requirements.txt
colorlog
ecdsa
oslo.i18n==1.5.0
stevedore==1.5.0
oslo.serialization==1.4.0
debtcollector==0.3.0
cliff==1.4.3
fabric>=1.8.2
ipaddr
jsondate
mysql-connector-python
oslo.config==2.3.0
oslo.log==1.12.0
pika==0.9.14
progressbar==2.3
pycrypto
python-cinderclient==1.1.0
python-glanceclient==0.12.0
python-keystoneclient==1.6.0
python-neutronclient==2.3.4
python-novaclient==2.29.0
python-swiftclient==2.3.1
pyyaml
redis
nose
sqlalchemy
# see https://github.com/paramiko/paramiko/issues/570 and https://github.com/paramiko/paramiko/pull/571
-e git://github.com/mirrorcoder/paramiko.git#egg=paramiko-1.16.0
